improved handling of null double values when creating csvs for entities in view models
removed enum for regular samplings and carrier boat landings and instead just used string constants: "rs" and "cbl"

added handling of null collection for carrierlandings when creating DBSummary when data is newly downloaded from the server

used string values for regular sampling and carrier boat landings when saving landing data from JSON to the local database

improved loading time of calendar view by not using the summaryview model as source and instead query from the database

+++ Entities/Database/WeightValidationUpdater.cs
@@ -48,7 +48,7 @@ namespace NSAP_ODK.Entities.Database
                 {
                     VesselCatchWV vesselCatchWV = new VesselCatchWV
                     {
-                        PK = c.VesselUnloadID,
+                        PK = (int)c.VesselUnloadID,
                         FromTotalCatch = c.FromTotalCatch,
                         Species_kg = c.Catch_kg,
                         Species_sample_kg = c.Sample_kg,
@@ -311,7 +311,7 @@ namespace NSAP_ODK.Entities.Database
                         {
                             VesselCatchWV vesselCatchWV = new VesselCatchWV
                             {
-                                PK = c.VesselUnloadID,
+                                PK = (int)c.VesselUnloadID,
                                 FromTotalCatch = c.FromTotalCatch,
                                 Species_kg = c.Catch_kg,
                                 Species_sample_kg = c.Sample_kg
